Explicitly define variable types for clarity
The type of the `solp` variable in the `foreach` loop has been explicitly defined as `Solp`. This change improves code readability and clarity by making the type of `solp` explicit.

The `tienePliego` variable has been changed from `var` to `bool` to explicitly define its type. This change ensures that the type of `tienePliego` is clear and unambiguous.

The `middleFileName` and `pdfFilename` variables have been changed from `var` to `string` to explicitly define their types. This change improves code readability and makes the types of these variables explicit.