ARMOA003-3450: Update conditions in HTML and service for display and PDF
Updated the condition in `listado-dashboard-comprador.component.html` to display a specific `<div>` element if `rowData.TipoSolpSap` is 2 or 3, or if `rowData.TipoSolp.Codigo` is 'CON_PLIEGO', or if `rowData.SolpConAdjuntos` is true. Added an `alt` attribute to the `<img>` element and bound a click event to `generarZipPliego` with `rowData.Id`.

Modified the condition in `ComprasService.cs` to generate a PDF file if `solp.TipoSolp.Codigo` is 'CON_PLIEGO', or if `solp.Urgencia` is true and `solp.TrabajoYaHecho` is not true, or if `solp.TipoSolpSap` is either `TipoSolpSap.Sap` or `TipoSolpSap.Mantenimiento`.