ARMOA003-3550: Update .sub-tabla width calculation in media query
Modified the CSS media query for screens with a min-width of 1240px.
Updated the .sub-tabla class's width from calc(100vw - 377px) to
calc(100vw - 220px). Added a fallback width using calc(100dvw - 220px)
to handle cases where dvw is not available. This change resolves a
conflict with the zoom: 90%; setting in MolinosGeneral.css.