ARMOA003-3282: Add permission check for editing "suplente" field
Introduced a new boolean property `tienePermisoEditarSuplente` in `UsuarioListComponent` to track user permissions for editing the "suplente" field. Added `verificarPermisosEdicion` method to set this property using `securityService`. Called this method in `ngOnInit` to initialize permissions on component load. Updated `FormControl` for `suplente` to be conditionally disabled based on `tienePermisoEditarSuplente`. Modified `valueChanges` subscription for `suplente` to conditionally enable/disable `fechaReasignar1` and reset it when disabled.