ARMOA003-3282: Add permission check before editing UsuarioReasignacion
Added a check to verify if the user has permission to edit a substitute before removing a `UsuarioReasignacion` object from the repository. If the user does not have the required permissions, an `UnauthorizedAccessException` is thrown with the message "No tiene permisos para editar el suplente". This ensures that only users with the appropriate permissions can perform this action.