Add RenspaExiste interface and SustentableRenspaExisteDto class
Added a new TypeScript interface `RenspaExiste` with boolean properties `RenspaExiste` and `MismoCuit` in `renspa-existe.interface.ts`. This file was included in the `SustitucionMOA.csproj` project file.

Introduced a new C# class `SustentableRenspaExisteDto` with boolean properties `RenspaExiste` and `MismoCuit` in the `SustitucionMOAModel.Dto` namespace. The `SustentableRenspaExisteDto.cs` file was added to the `SustitucionMOAModel.csproj` project file.