Set default dates for Pliego and update modification date
The `FechaAlta` property in the `Pliego` class within the `SustitucionMOAModel.Entities` namespace has been modified to have a default value of `DateTime.Now`. This means that when a new instance of the `Pliego` class is created, `FechaAlta` will automatically be set to the current date and time.

In the `ComprasService` class within the `SustitucionMOAUtils.Services` namespace, a new line has been added to set the `FechaModificacion` property of the `pliegoEntity` object to the current date and time (`DateTime.Now`). This ensures that whenever modifications are made to the `pliegoEntity`, the `FechaModificacion` property is updated to reflect the time of the modification.