Add PliegoDto class and update project file
Added PliegoDto.cs defining the PliegoDto class in the SustitucionMOAModel.Dto.PliegoMultiple namespace. This class includes properties like Id, Codigo, NombreObra, FiscalContrato, Email, Telefono, FechaHoraEntrega, and more. It also provides methods for converting from a Pliego object to a PliegoDto object and an explicit operator for this conversion. Updated SustitucionMOAModel.csproj to include the new file in the project compilation.