Add JoinPipe to project and shared module
A new TypeScript file `join.pipe.ts` has been added, defining an Angular pipe named `JoinPipe`. This pipe takes an array of strings and a separator as input and returns a single string where the array elements are joined by the specified separator. The `JoinPipe` has been imported and added to the `declaredAndExported` array in the `shared.module.ts` file, making it available for use throughout the module. The addition is also reflected in the `SustitucionMOA.csproj` file.