Add Solps property to Pliego class with InverseProperty
Introduced a new navigation property `Solps` to the `Pliego` class, which is a virtual collection of `Solp` entities. This property is decorated with the `InverseProperty` attribute, establishing a bidirectional relationship with the `Pliego` property in the `Solp` class. This enhancement improves navigation and management of related entities within the Entity Framework.