Add GetSolpDisponiblesPliegosMultiple method
A new method GetSolpDisponiblesPliegosMultiple has been added to the IPliegoMultipleService interface. This method filters and retrieves a list of Solp entities based on specific criteria, such as EstadoSolpSap.CodigoSap, TipoSolp.Codigo, and Posiciones.TipoPosicion.Codigo. The method ensures that the Solp entities do not have certain boolean flags set to true, are not associated with a multiple Pliego, and do not have any Posiciones with adjudications. The filtered Solp entities are then converted to SolpDto objects and returned as a list.