Add conditional params and improve initial search trigger
In `pliegoMultiple.service.ts`, added new parameters (`fechaInicio`,
`fechaFin`, `creador`, `fiscal`, `sap`, and `mantenimiento`) to the
`HttpParams` object conditionally, allowing for more flexible querying.

In `vincular-solp-pliego-multiple.component.ts`, replaced the initial
call to `this.getSolps()` with `this.debouncer.next();` to ensure the
debounce mechanism is used consistently, even for the initial search.