Update SolpDto properties to use new source properties
Updated the `Creador` property in the `SolpDto` class to use the
`Mail` property of `solp.UsuarioCreacion` instead of the
`ObtenerRazonSocial()` method. Updated the `Fiscal` property in
the `SolpDto` class to use the `Email` property of `solp.Pliego`
instead of the `FiscalContrato` property. These changes reflect
an update in the data model or a shift in the required information
for these properties.