Add descargarZipPliego method to PliegoMultipleService
Added a new method `descargarZipPliego` to the `PliegoMultipleService` class.
This method takes an `idPliego` parameter of type `number` and returns an `Observable<any>`.
It creates an `HttpParams` object, sets the `pliegoId` parameter with the provided `idPliego` value, and makes an HTTP GET request to the `/api/PliegoMultiple/DescargarZipPliego` endpoint.
The request includes the `params` and `headers` properties.