Add TipoPliego property to ComprasController
Introduced a new property `TipoPliego` in `ComprasController.cs` within the `SustitucionMOA.Controllers` namespace. This property is a list of objects, each containing an `Id` and a `Descripcion`. The list includes two items: one with `Id` 0 and `Descripcion` "Pliego única SOLP", and another with `Id` 1 and `Descripcion` "Pliego múltiple SOLP".