Add feature to copy pliego with new route and component mode
Introduced a new route in `compras-routing.module.ts` to handle the path `solp/:id/:tipoSolp/:action` with the `SolpComponent`. Updated `dashboard-pliego-multiple.component.html` to add a click event handler for copying a pliego. Added a new method `copiarPliego` in `dashboard-pliego-multiple.component.ts` to handle the copy action and navigate to the new route.

Modified `solp.component.ts` to handle the new `action` parameter in the route, specifically for the copy action. Introduced a new `ComponentMode` called `Copy` in `solp.component.ts` and `solp.utils.ts`. Updated the `ngOnInit` method in `solp.component.ts` to set the component mode to `Copy` if the `action` parameter is 'copy'. Added a new method `traierPliegoMultipleId` in `solp.component.ts` to handle fetching data for pliego multiple by ID. Added a new getter `esCopiaPliegoMultiple` in `solp.component.ts` to check if the current mode is `Copy` and the type is `PLIEGO_MULTIPLE`. Updated the logic in `solp.component.ts` to call `traierPliegoMultipleId` instead of `traerSolpId` when the type is `PLIEGO_MULTIPLE`.

Added a `debugger` statement in `solp.component.ts` for debugging purposes. Updated the `ComponentMode` enum in `solp.utils.ts` to include the new `Copy` mode.