Add pliegoId parameter to getSolpDisponiblesPliegosMultiple
The getSolpDisponiblesPliegosMultiple method now accepts an
additional parameter pliegoId of type number | null. The method
parameters have been reformatted for better readability. A new
condition has been added to set the pliegoId parameter in the
HttpParams object if it is provided. The HTTP GET request now
includes the pliegoId parameter if it is present.