Add pliegoId input and default center setup
Introduce a new input property `[pliegoId]="pliegoIdWhenEditing"` to the `<vincular-solp-pliego-multiple>` component in `solp.component.html`. This property is provided by a new getter method `pliegoIdWhenEditing` in `SolpComponent`, which returns `null` or `solpId` based on the editing mode. Additionally, add a private method `setupCentroPorDefecto` to set the `centroPorDefecto` property of `solpActual` to a default value if a match is found in the `combos.Centro` array with the `Codigo` "1029".