Expand if condition to include esCopiaPliegoMultiple
The condition in the `if` statement has been modified to include an additional check for `this.esCopiaPliegoMultiple`. Previously, the code block inside the `if` statement would execute only if `this.esEdicionSolp` was true. Now, it will execute if either `this.esEdicionSolp` or `this.esCopiaPliegoMultiple` is true, allowing for additional scenarios to be handled.