Add condition for copying multiple "pliego" in finalizar
Extended the conditional check in the `finalizar` method of the
`SolpComponent` class to include `esCopiaPliegoMultiple`. Now,
the method will call `guardarCambios` and `guardarPliegoMultiple`
if any of `esCreacionPliegoMultiple`, `esEdicionPliegoMultiple`,
or `esCopiaPliegoMultiple` are true. This change ensures that
the same actions are performed when copying multiple "pliego"
as when creating or editing them.