Add check for esCopiaPliegoMultiple in if condition
The condition in the `if` statement within the `SolpComponent` class has been updated to include an additional check for `this.esCopiaPliegoMultiple`. Previously, the condition only checked for `this.esCreacionPliegoMultiple` or `this.esEdicionPliegoMultiple`. Now, the block of code within the `if` statement will execute if any of these three conditions are met.