Add TipoPliego parameter to ComprasService methods
Introduce a new parameter `TipoPliego tipoPliego` with a default value of `TipoPliego.All` to method signatures in `IComprasService.cs` and `ComprasService.cs`. Update the `filtroServicio` expression in `ComprasService.cs` to include conditions that filter based on the value of `tipoPliego`, allowing filtering of positions based on whether they are associated with a single or multiple pliego.