fix Pliego multiple
Add conditional filter logic and read technical specs.
Introduce a conditional block to handle `incluirGuardadas` flag,
resetting the filters list if true before adding a new filter.
Also, read and assign the content of the file specified in
`pliegoReturn.EspecificacionesTecnicas` to ensure technical
specifications are included in the `pliegoReturn` object.