MOA-450
Add conditional disabling to checkbox in contrato form

The code change modifies the <input type="checkbox"> element within a <label class="switch"> in the crear-contrato.fijacion.component.html file. Specifically, it adds a [disabled]="isVisiblePizarra()" attribute to the checkbox input. This ensures that the checkbox for contrato.Pizarra will be disabled if the isVisiblePizarra() function returns true, enhancing the form's interactivity and user experience by preventing changes to the checkbox when certain conditions are met.